Monkey pole Jump problem( leetcode Jump array Problem no. 55)
In a circus a monkey is jumping over different poles .Each pole is having a number the ring master make sure that the monkey jump to the pole according to the number given to it. If Monkey starts from 1st pole then in how many jumps it will reach till the end when the numbers of poles are:--
1, 3, 5, 8, 9, 2, 6, 7, 6, 8, 9.
Note:-- 
First jump from 1st element to 2nd pole with value 3. Now, from here monkey jump to 5th pole with value 9, and from here monkey will jump to the last pole.So in total 3 jumps.
output-- 3